+++ frontend/srcs/src/app/app.module.ts
@@ -7,7 +7,6 @@ import { AdministrationComponent,
          ChatComponent,
          FriendListComponent,
          GameComponent,
-         HistoryComponent,
          HomeComponent,
          LoginModule,
          NavbarComponent,
@@ -44,7 +43,8 @@ import { AvatarComponent,
          AdminstrationChannelsComponent,
          AdministrationChannelManagementUsersComponent,
          AdministrationChannelManagementActionsComponent,
-         AdministrationUsersManagementActionsComponent } from './components';
+         AdministrationUsersManagementActionsComponent, 
+        AdministrationChannelWindowComponent} from './components';
 import { DateMutations } from './utils';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { TagModule } from 'primeng/tag';
@@ -56,10 +56,8 @@ import { ProfileAvatarComponent } from './components/avatar/profile-avatar/profi
 import { TooltipModule } from 'primeng/tooltip';
 import { RankTableModule } from './modules/rank-table/rank-table.module'
 import { TruncatePipe } from './pipes/truncate.pipe';
-//import { PongTVComponent } from './modules/pongtv/pongtv.component';
 import { ChannelAvatarEditorComponent } from './components/avatar/channel-avatar-editor/channel-avatar-editor.component';
 import { NgbDatepickerModule } from '@ng-bootstrap/ng-bootstrap';
-//import { PongTVComponent } from './modules/pongtv/pongtv.component';
 
 const config: SocketIoConfig = { url: environment.apiUrl };
 
@@ -97,6 +95,7 @@ const config: SocketIoConfig = { url: environment.apiUrl };
     AdministrationChannelManagementUsersComponent,
     AdministrationChannelManagementActionsComponent,
     AdministrationUsersManagementActionsComponent,
+    AdministrationChannelWindowComponent,
     FriendListComponent,
     ChatSidebarComponent,
     FriendSidebarComponent,