+++ frontend/srcs/src/app/components/administration/index.ts
@@ -3,3 +3,4 @@ export * from './administration-channels-list/administration-channels-list.compo
 export * from './administration-channel-management-users/administration-channel-management-users.component'
 export * from './administration-channel-management-actions/administration-channel-management-actions.component'
 export * from './administration-users-management-actions/administration-users-management-actions.component'
+export * from './administration-channel-window/administration-channel-window.component'